Clear outputs

+++ data-viz/tutorial/exercises/fetch_data.py
@@ -8,8 +8,14 @@ import zipfile
 
 URL = "https://earthquake.usgs.gov/fdsnws/event/1/query.csv?starttime={start}&endtime={end}&minmagnitude=2.0&orderby=time"
 
+data_path = '../../data'
 def get_earthquake_data():
-    for yr in range(2000, 2001):
+    earthquakes_parquet_file = f'{data_path}/earthquakes.parq'
+    if os.path.isfile(earthquakes_parquet_file):
+        print('Earthquakes dataset present, skipping download')
+        return
+
+    for yr in range(2000, 2010):
         for m in range(1, 13):
             if os.path.isfile('{yr}_{m}.csv'.format(yr=yr, m=m)):
                 continue
@@ -20,7 +26,7 @@ def get_earthquake_data():
                 f.write(requests.get(URL.format(start=start, end=end)).content.decode('utf-8'))
 
     dfs = []
-    for i in range(2000, 2001):
+    for i in range(2000, 2010):
         for m in range(1, 13):
             if not os.path.isfile('%d_%d.csv' % (i, m)):
                 continue
@@ -36,9 +42,8 @@ def get_earthquake_data():
             os.remove(filePath)
         except:
             print("Error while deleting file : ", filePath)
-
     df = pd.concat(dfs, sort=True)
-    df.to_parquet('../../data/earthquakes.parq', 'fastparquet')
+    df.to_parquet(earthquakes_parquet_file, 'fastparquet')
 
 def download_url(url, save_path, chunk_size=128):
     r = requests.get(url, stream=True)
@@ -48,7 +53,10 @@ def download_url(url, save_path, chunk_size=128):
             
 def get_population_data():
     fname = 'gpw_v4_population_density_rev11_2010_2pt5_min.zip'
+    if os.path.isfile(f'{data_path}/{fname}'):
+        print('Population dataset present, skipping download')
+        return    
     download_url(f'https://earth-data.s3.amazonaws.com/{fname}', fname)
     with zipfile.ZipFile(fname, 'r') as zip_ref:
-        zip_ref.extractall('../../data/')
+        zip_ref.extractall(data_path)
     os.remove(fname)
